V1 of factored-out ManaSource. Some features not yet implemented! Also finished make/linux.bash

+++ make/linux.bash
@@ -1,15 +1,16 @@
 # Run from magnus root directory.
 cd $GOPATH/src/github.com/runningwild/magnus
-go build --tags host .
 
-rm -rf bin
-mkdir -p bin/fail
+if go build --tags host . ; then
+		rm -rf bin
+		mkdir -p bin/fail
 
-cp magnus bin/fail
-cp ../glop/gos/linux/lib/libglop.so bin/fail/libglop.so
-echo "LD_LIBRARY_PATH=$LD_LIBRARY_PATH:. ./magnus" > bin/fail/runme
-chmod 776 bin/fail/runme
-cp -r data/* bin/
+		cp magnus bin/fail
+		cp ../glop/gos/linux/lib/libglop.so bin/fail/libglop.so
+		echo "LD_LIBRARY_PATH=$LD_LIBRARY_PATH:. ./magnus" > bin/fail/runme
+		chmod 776 bin/fail/runme
+		cp -r data/* bin/
 
-cd bin/fail
-./runme
+		cd bin/fail
+		./runme
+fi
